Add retract; merge api/user[s] files
