+++ web/template/test.gohtml
@@ -15,7 +15,7 @@
     <title>test</title>
 </head>
 <body>
-<div x-data="watch.chatContext(12845, {ID: 1, name: 'admin'})">
+<div x-data="watch.chatContext(12845, {ID: 1, name: 'admin', isAdmin: true})">
     <div x-data="watch.chatPromptContext(12845, 1)"
          style="padding: 16px; margin: 8px; border: 3px solid blue;">
         <div>
